requirements.txt
cassandra-driver
kafka-python
requests